Refactor out `TimerSourceActionBase`
The `Interact` action had a lot of logic in it for configuring which
timer source to interact with.  This logic will be useful for other
actions that interact with specific timer sources, so this patch moves
it out into its own base class.
